backend/requirements.txt
fastapi
uvicorn
pydantic
gpiozero
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-ssd1306
Pillow
